Mejoradas todas las funcionalidades de proyectos, listas y tareas en el cliente para tratar con otros usuarios

+++ cliente/main.go
@@ -66,14 +66,18 @@ func main() {
 	})
 
 	//Cambiar a la pantalla de añadir una lista
-	UI.Bind("changeToAddListGO", func(proyectID string, proyectName string) bool {
-		//Recupero los usuarios del proyecto
-		usersProyect, tokenExpire := models.GetUsersProyect(proyectID)
-		if tokenExpire {
-			return tokenExpire
+	UI.Bind("changeToAddListGO", func(proyectID string, proyectName string) []bool {
+
+		if !utils.CheckExpirationTimeToken(models.UserSesion.Token) {
+			return []bool{false, true}
+		} else {
+			if models.ExistProyect(proyectID) {
+				//Recupero los usuarios del proyecto
+				usersProyect, _ := models.GetUsersProyect(proyectID)
+				ChangeViewAddList(config.PreView+"addList.html", proyectID, proyectName, usersProyect)
+			}
+			return []bool{false, false}
 		}
-		ChangeViewAddList(config.PreView+"addList.html", proyectID, proyectName, usersProyect)
-		return tokenExpire
 	})
 
 	//Cambiar a la pestaña de ver las tareas de una lista dado el ID de la lista
@@ -83,7 +87,9 @@ func main() {
 		} else {
 			//Compruebo que la lista existe
 			if models.ExistList(listID) {
-				tasks := models.GetTasksByList(listID)
+				tasks, tasksCipher := models.GetTasksByList(listID)
+				models.TasksLocal = tasksCipher
+
 				ChangeViewTasks(config.PreView+"tasks.html", tasks, listID, nil, listName)
 			}
 			return []bool{false, false}
@@ -92,23 +98,19 @@ func main() {
 
 	//REVISAR SI QUIERO TRAERME DEL SERVIDOR O DE LOCAL
 	//Cambiar a la pestaña de ver la configuracion del proyecto dado su ID
-	UI.Bind("changeToProyectConfigGO", func(proyectID string) bool {
+	UI.Bind("changeToProyectConfigGO", func(proyectID string) []bool {
 		if !utils.CheckExpirationTimeToken(models.UserSesion.Token) {
-			return true
+			return []bool{false, true}
 		} else {
-			proyectCipher := models.GetProyect(proyectID)
-			proyectKey := models.GetProyectKey(proyectID, models.UserSesion.Email)
-			proyect := models.DescifrarProyecto(proyectCipher, proyectKey)
-
-			// var proyect models.Proyect
-			// for i := 0; i < len(models.DatosUsuario); i++ {
-			// 	if models.DatosUsuario[i].Proyecto.ID == proyectID {
-			// 		proyect = models.DatosUsuario[i].Proyecto
-			// 	}
-			// }
-			emailsNotInProyect := models.GetEmailsNotInProyect(proyect)
-			ChangeViewConfigProyect(config.PreView+"configProyect.html", proyect, emailsNotInProyect, models.UserSesion.Email)
-			return false
+			if models.ExistProyect(proyectID) {
+				proyectKey := models.GetProyectKey(proyectID, models.UserSesion.Email)
+				proyectCipher := models.GetProyect(proyectID)
+				proyect := models.DescifrarProyecto(proyectCipher, proyectKey)
+				emailsNotInProyect := models.GetEmailsNotInProyect(proyect)
+				ChangeViewConfigProyect(config.PreView+"configProyect.html", proyect, emailsNotInProyect, models.UserSesion.Email)
+				return []bool{true, false}
+			}
+			return []bool{false, false}
 		}
 	})
 
@@ -130,9 +132,10 @@ func main() {
 			// 		}
 			// 	}
 			// }
-			listCipher := models.GetList(listID)
-			if !listCipher.ID.IsZero() {
-				listKey := models.GetListKey(listID)
+			listKey := models.GetListKey(listID)
+			if len(listKey) > 0 {
+				listCipher := models.GetList(listID)
+
 				list := models.DescifrarLista(listCipher, listKey)
 
 				//Elimino los usuarios que ya pertenecen a la lista
@@ -150,15 +153,18 @@ func main() {
 	})
 
 	//Cambiar a la pestaña de añadir Tareas
-	UI.Bind("changeToAddTaskGO", func(listID string) bool {
+	UI.Bind("changeToAddTaskGO", func(listID string) []bool {
 		if !utils.CheckExpirationTimeToken(models.UserSesion.Token) {
-			return true
+			return []bool{false, true}
 		} else {
-			listCipher := models.GetList(listID)
-			list := models.DescifrarLista(listCipher, models.GetListKey(listID))
-			listUsers := list.Users
-			ChangeViewTasks(config.PreView+"addTask.html", nil, listID, listUsers, list.Name)
-			return false
+			if models.ExistList(listID) {
+				listCipher := models.GetList(listID)
+				list := models.DescifrarLista(listCipher, models.GetListKey(listID))
+				listUsers := list.Users
+				ChangeViewTasks(config.PreView+"addTask.html", nil, listID, listUsers, list.Name)
+				return []bool{true, false}
+			}
+			return []bool{false, false}
 		}
 	})
 
@@ -166,9 +172,10 @@ func main() {
 	UI.Bind("changeToTaskConfigGO", func(taskID string, listID string) []bool {
 		if !utils.CheckExpirationTimeToken(models.UserSesion.Token) {
 			return []bool{false, true}
-		} else {
+		} else if models.ExistList(listID) {
 			return []bool{LoadTask(taskID, listID), false}
 		}
+		return []bool{false, false, false}
 	})
 
 	/*------------------------------------------------------------------------- Binds Funcionalidades  ------------------------------------------------------------------*/
@@ -224,48 +231,57 @@ func main() {
 
 	//Eliminar un proyecto
 	UI.Bind("deleteProyectGO", func(proyectID string) []bool {
-		deleteOk, tokenExpire := models.DeleteProyect(proyectID)
-		return []bool{deleteOk, tokenExpire}
+		if !utils.CheckExpirationTimeToken(models.UserSesion.Token) {
+			return []bool{false, false, true}
+		} else if models.ExistProyect(proyectID) {
+			deleteOk, tokenExpire := models.DeleteProyect(proyectID)
+			return []bool{true, deleteOk, tokenExpire}
+		}
+		return []bool{false, false, false}
 	})
 
 	//Borrar un usuario de un Proyecto
 	UI.Bind("deleteUserProyectGO", func(userEmail string, proyectID string) []bool {
 		if !utils.CheckExpirationTimeToken(models.UserSesion.Token) {
-			return []bool{false, true}
-		} else {
+			return []bool{false, false, true}
+		} else if models.ExistProyect(proyectID) {
 			deleteOk, tokenExpire := models.DeleteUserProyect(proyectID, userEmail)
-			return []bool{deleteOk, tokenExpire}
+			return []bool{true, deleteOk, tokenExpire}
 		}
+		return []bool{false, false, false}
 	})
 
 	//Añadir un usuario al proyecto
 	UI.Bind("addUserProyectGO", func(userEmail string, proyectID string) []bool {
 		if !utils.CheckExpirationTimeToken(models.UserSesion.Token) {
-			return []bool{false, true}
-		} else {
+			return []bool{false, false, true}
+		} else if models.ExistProyect(proyectID) {
 			addOk, tokenExpire := models.AddUserProyect(proyectID, userEmail)
-			return []bool{addOk, tokenExpire}
+			return []bool{true, addOk, tokenExpire}
 		}
+		return []bool{false, false, false}
 	})
 
 	//Actualizar un proyecto
 	UI.Bind("updateProyectGO", func(newProyect models.Proyect) []bool {
 		if !utils.CheckExpirationTimeToken(models.UserSesion.Token) {
-			return []bool{false, true}
-		} else {
+			return []bool{false, false, true}
+		} else if models.ExistProyect(newProyect.ID) {
 			updateOk, tokenExpire := models.UpdateProyect(newProyect)
-			return []bool{updateOk, tokenExpire}
+			return []bool{true, updateOk, tokenExpire}
 		}
+		return []bool{false, false, false}
 	})
 
 	//Añadir una lista
 	UI.Bind("addListGO", func(list models.List, proyectID string) []bool {
 		if !utils.CheckExpirationTimeToken(models.UserSesion.Token) {
-			return []bool{false, true}
-		} else {
+			return []bool{false, false, true}
+		} else if models.ExistProyect(proyectID) {
 			addOk, tokenExpire := models.CreateList(list, proyectID)
-			return []bool{addOk, tokenExpire}
+			return []bool{true, addOk, tokenExpire}
 		}
+		return []bool{false, false, false}
 	})
 
 	//Eliminar una lista
@@ -273,28 +289,34 @@ func main() {
 		if !utils.CheckExpirationTimeToken(models.UserSesion.Token) {
 			return []bool{false, true}
 		} else {
-			//Consigo los usuarios de la lista
-			usersList := models.GetUsersList(listID)
-			//Elimino la lista de todas las relaciones
-			for i := 0; i < len(usersList); i++ {
-				models.DeleteRelationList(proyectID, listID, usersList[i])
+			if models.ExistList(listID) {
+				//Consigo los usuarios de la lista
+				usersList := models.GetUsersList(listID)
+				//Elimino la lista de todas las relaciones
+				for i := 0; i < len(usersList); i++ {
+					models.DeleteRelationList(proyectID, listID, usersList[i])
+				}
+				//Elimino la lista
+				deleteOk, tokenExpire := models.DeleteList(listID)
+				return []bool{deleteOk, tokenExpire}
 			}
-			//Elimino la lista
-			deleteOk, tokenExpire := models.DeleteList(listID)
-			return []bool{deleteOk, tokenExpire}
+			return []bool{false, false}
 		}
 	})
 
 	//Borrar un usuario de una Lista
 	UI.Bind("deleteUserListGO", func(userEmail string, listID string, proyectID string) []bool {
 		if !utils.CheckExpirationTimeToken(models.UserSesion.Token) {
-			return []bool{false, true}
-		} else {
+			return []bool{false, false, true}
+		} else if models.ExistList(listID) {
+
 			//Borro al usuario de la lista
 			models.DeleteUserList(listID, userEmail)
 			//Borro la lista en la relacion donde aparece
 			deleteOk := models.DeleteRelationList(proyectID, listID, userEmail)
-			return []bool{deleteOk, false}
+			return []bool{true, deleteOk, false}
+		} else {
+			return []bool{false, false, false}
 		}
 	})
 
@@ -303,18 +325,21 @@ func main() {
 		if !utils.CheckExpirationTimeToken(models.UserSesion.Token) {
 			return []string{"false", "true"}
 		} else {
-			updateStatus, tokenExpire := models.UpdateList(newList)
-			return []string{updateStatus, tokenExpire}
+
+			listOK, updateStatus, tokenExpire := models.UpdateList(newList)
+			return []string{listOK, updateStatus, tokenExpire}
 		}
 	})
 
 	//Añadir un usuario a una lista
 	UI.Bind("addUserListGO", func(userEmail string, proyectID string, listID string) []bool {
 		if !utils.CheckExpirationTimeToken(models.UserSesion.Token) {
-			return []bool{false, true}
-		} else {
+			return []bool{false, false, true}
+		} else if models.ExistList(listID) {
 			addOK := models.AddUserList(userEmail, proyectID, listID)
-			return []bool{addOK, false}
+			return []bool{true, addOK, false}
+		} else {
+			return []bool{false, false, false}
 		}
 	})
 
@@ -338,8 +363,7 @@ func main() {
 			addOK, tokenExpire := models.DeleteTask(taskID)
 			return []bool{true, addOK, tokenExpire}
 		}
-		addOK, tokenExpire := models.DeleteTask(taskID)
-		return []bool{false, addOK, tokenExpire}
+		return []bool{false, false, false}
 	})
 
 	//Actualizar una tarea
@@ -353,6 +377,14 @@ func main() {
 			}
 			return []string{"false", "false", "false"}
 		}
+	})
+
+	UI.Bind("checkChangesGO", func() bool {
+		return models.CheckChanges()
+	})
+
+	UI.Bind("checkTaskChangesGO", func(listID string) bool {
+		return models.CheckTaskChanges(listID)
 	})
 
 	sigc := make(chan os.Signal)
